:art: [feat] 상대방 정보 페이지 연결
