+++ src/app/(route)/chat/partner-info/[slug]/layout.tsx
@@ -1,3 +1,10 @@
+import type { Metadata } from 'next'
+
+export const metadata: Metadata = {
+  title: 'Create Next App',
+  description: 'Generated by create next app',
+}
+
 export default function PartnerInfoLayout({
   children,
 }: {
