--- src/app/(route)/partnerinfo/layout.tsx
@@ -1,14 +1,0 @@
-import type { Metadata } from 'next'
-
-export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-}
-
-export default function PartnerInfo({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return <main>{children}</main>
-}
